% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demine.R
\name{afficher_grille}
\alias{afficher_grille}
\title{Show the grid of the game with the flags and the revealed cells.}
\usage{
afficher_grille(grille, visible, drapeaux)
}
\arguments{
\item{grille}{A matrix of integers. The value of each cell is the number of mines adjacent to it or a bomb itself.}

\item{visible}{A matrix of booleans. The value of each cell is TRUE if the cell is revealed, FALSE otherwise.}

\item{drapeaux}{A matrix of booleans. The value of each cell is TRUE if a flag is placed on it, FALSE otherwise.}
}
\value{
A visual representation of the grid.
}
\description{
Show the grid of the game with the flags and the revealed cells.
}
\examples{

afficher_grille(matrix(0, nrow = 2, ncol = 2), matrix(FALSE, nrow = 2, ncol = 2), matrix(c(TRUE,FALSE,FALSE,FALSE), nrow = 2))
afficher_grille(matrix(c(-1,1,0,1,1,0), nrow = 2), matrix(c(FALSE,FALSE,TRUE,TRUE,FALSE,FALSE), nrow = 2), matrix(FALSE, nrow = 2, ncol = 3))

}
