% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demine.R
\name{calcul_points}
\alias{calcul_points}
\title{Count the points the player do each move.
The player gets points for each move. The faster he is, the more points he gets. Even more if he reveals a lot of cells at once.}
\usage{
calcul_points(points, temps_depart, nb_cases, nb_nouvelles_cases)
}
\arguments{
\item{points}{The number of points the player currently has.}

\item{temps_depart}{The time at which the game started.}

\item{nb_cases}{The number of cases the player revealed before the move.}

\item{nb_nouvelles_cases}{The number of cases the player revealed after the move.}
}
\value{
The number of points the player has after the move.
}
\description{
Count the points the player do each move.
The player gets points for each move. The faster he is, the more points he gets. Even more if he reveals a lot of cells at once.
}
