% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demine.R
\name{calculer_mines_adjacentes}
\alias{calculer_mines_adjacentes}
\title{Calculate the number of mines adjacent to a cell.
We don't care it the cell itself is a mine or not, we just don't use it if it is.}
\usage{
calculer_mines_adjacentes(grille, i, j)
}
\arguments{
\item{grille}{A matrix of integers. The value of each cell is the number of mines adjacent to it or a bomb itself.}

\item{i}{The row of the cell.}

\item{j}{The column of the cell.}
}
\value{
The number of mines adjacent to the cell.
}
\description{
Calculate the number of mines adjacent to a cell.
We don't care it the cell itself is a mine or not, we just don't use it if it is.
}
\examples{

calculer_mines_adjacentes(matrix(c(0,-1,-1,-1), nrow = 2), 1, 1)

}
