% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_grid}
\alias{convert_grid}
\title{Update the grid}
\usage{
convert_grid(grid, rows, columns)
}
\arguments{
\item{grid}{The grid to update}

\item{rows}{The number of rows}

\item{columns}{The number of columns}
}
\value{
The updated grid
}
\description{
Update the grid
}
