% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_buttons}
\alias{generate_buttons}
\title{Prepares the grid for the pictures}
\usage{
generate_buttons(grid_values, grid_hidden, grid_flags, rows, columns)
}
\arguments{
\item{grid_values}{The grid of values}

\item{grid_hidden}{The grid of hidden cells}

\item{grid_flags}{The grid of flags}

\item{rows}{The number of rows}

\item{columns}{The number of columns}
}
\value{
The grid of buttons
}
\description{
Prepares the grid for the pictures
}
