% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_grid}
\alias{generate_grid}
\title{Generate the grid with the mines}
\usage{
generate_grid(grid, rows, columns, bombs)
}
\arguments{
\item{grid}{The grid}

\item{rows}{The number of rows}

\item{columns}{The number of columns}

\item{bombs}{The number of bombs}
}
\value{
The grid with the mines
}
\description{
Generate the grid with the mines
}
