% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{init_grid}
\alias{init_grid}
\title{Initialize the grid}
\usage{
init_grid(rows, columns, value)
}
\arguments{
\item{rows}{The number of rows}

\item{columns}{The number of columns}

\item{value}{The value to fill the grid with}
}
\value{
A matrix of the specified size filled with the specified value
}
\description{
Initialize the grid
}
