% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demine.R
\name{jouer_partie}
\alias{jouer_partie}
\title{Play a game of minesweeper.}
\usage{
jouer_partie()
}
\value{
Each time the player makes a move, the function prints the new state of the game. At the end of the game, the function prints if he lost or won and his number of points.
}
\description{
Play a game of minesweeper.
}
