% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reveal_block}
\alias{reveal_block}
\title{Used to reveal adjactent cells that don't have bomb either after you clicked
on a cell from a block of cells}
\usage{
reveal_block(grille, visible, i, j)
}
\arguments{
\item{grille}{Grid of bombs/numbers}

\item{visible}{Grid of cells that have been revealed by the click already (empty at first)}
}
\value{
visible
}
\description{
Used to reveal adjactent cells that don't have bomb either after you clicked
on a cell from a block of cells
}
