% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-module.R
\name{time}
\alias{time}
\title{Time module}
\usage{
time(input, output, session, start = reactive(0))
}
\arguments{
\item{start}{A reactive value that starts the timer}
}
\value{
A reactive value with the time in seconds
}
\description{
Time module
}
