% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{to_coordinates}
\alias{to_coordinates}
\title{Used to get the coordinates of the cells that have been revealed
by reveal_block.}
\usage{
to_coordinates(matrice)
}
\arguments{
\item{matrice}{should be the matrix "visible" returned by reveal_block}
}
\value{
a matrix with the coordinates of the cells that have been revealed
}
\description{
Used to get the coordinates of the cells that have been revealed
by reveal_block.
}
