% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win-module.R
\name{win}
\alias{win}
\title{Victory pop-up}
\usage{
win(input, output, session, score, time)
}
\arguments{
\item{score}{The score}

\item{time}{The time}
}
\value{
A reactive value that refreshes the page
}
\description{
Victory pop-up
}
