% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win-module.R
\name{win_UI}
\alias{win_UI}
\title{Victory pop-up}
\usage{
win_UI(id, score, time)
}
\arguments{
\item{id}{The module id}

\item{score}{The score}

\item{time}{The time}
}
\value{
A modal dialog
}
\description{
Victory pop-up
}
