% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_button}
\alias{update_button}
\title{Used to update the picture on a button after it's been clicked}
\usage{
update_button(i, j, grid_values)
}
\arguments{
\item{i}{The row}

\item{j}{The column}

\item{grid_values}{The grid of values}
}
\description{
Used to update the picture on a button after it's been clicked
}
